cleaned up the write loop
